+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/RealmInstant.kt
@@ -41,12 +41,12 @@ public interface RealmInstant : Comparable<RealmInstant> {
         /**
          * Minimum timestamp that can be stored in Realm.
          */
-        public val MIN: RealmInstant = fromEpochSeconds(Long.MIN_VALUE, -999_999_999)
+        public val MIN: RealmInstant = from(Long.MIN_VALUE, -999_999_999)
 
         /**
          * Maximum timestamp that can be stored in Realm.
          */
-        public val MAX: RealmInstant = fromEpochSeconds(Long.MAX_VALUE, 999_999_999)
+        public val MAX: RealmInstant = from(Long.MAX_VALUE, 999_999_999)
 
         /**
          * Creates a [RealmInstant] that is the [epochSeconds] number of seconds from the UNIX epoch
@@ -63,7 +63,7 @@ public interface RealmInstant : Comparable<RealmInstant> {
          * If the timestamp exceed the maximal bounds of [epochSeconds], the Timestamp will clamp
          * to either [MIN] or [MAX].
          */
-        public fun fromEpochSeconds(epochSeconds: Long, nanosecondAdjustment: Int): RealmInstant {
+        public fun from(epochSeconds: Long, nanosecondAdjustment: Int): RealmInstant {
             val secAdjustment: Long = (nanosecondAdjustment / SEC_AS_NANOSECOND).toLong()
             val nsAdjustment: Int = nanosecondAdjustment % SEC_AS_NANOSECOND
             var s: Long = epochSeconds + secAdjustment